Add prev slide and table support
A new keybinding has been introduced namely `p` for previous slides.

Also, the main formatter has been switched to remarkable as it enjoys
better support and is fast as well. The table support has been added and
bugs in the partitioner have been removed. specifically the erroneous
setting of the slides variable.