Add Settings, Partitioner Limit and Design Changes
Settings Component has been created to handle all the changes that can
be done at the start of the presentation. The link between the settings
and the global state has not been made. It is planned to use redux for
settings and file management.

Partitiner now has a configurable function parameter to set the maximum
number of paragraphs that can be clumped together in one slide.

Some tweaks have been made on the sass file, includeing a change to
Aileron for content font.