Add density function and tests
Added the density function which numerically sets the value of a piece
of markdown. This will be later used to divide the slides. We shall try
to keep the density in each slides equal and set it so that the
visibility of the text is not compromised.

I have also added tests. The tests are simple and check whether the
density functions are working correctly.